Run pylint on tests and fix warnings
