+++ tests/unit_tests/repobee_plug/test_serialize.py
@@ -1,5 +1,6 @@
-import pytest
 import collections
+
+import pytest
 
 import repobee_plug as plug
